v0.4.0 Added support to magazines Make and RollingStone.

+++ src/main/java/com/newsup/kernel/Date.java
@@ -7,12 +7,12 @@ import java.util.TimeZone;
 
 public class Date {
 
-    private static final int SECOND_TIME = 1000;
-    private static final int MINUTE_TIME = 60 * SECOND_TIME;
-    private static final int HOUR_TIME = 60 * MINUTE_TIME;
-    private static final int DAY_TIME = 24 * HOUR_TIME;
-    private static final int MONTH_TIME = 30 * DAY_TIME;
-    private static final int YEAR_TIME = 365 * DAY_TIME;
+    private static final long SECOND_TIME = 1000;
+    private static final long MINUTE_TIME = 60 * SECOND_TIME;
+    private static final long HOUR_TIME = 60 * MINUTE_TIME;
+    private static final long DAY_TIME = 24 * HOUR_TIME;
+    private static final long MONTH_TIME = 30 * DAY_TIME;
+    private static final long YEAR_TIME = 365 * DAY_TIME;
 
     String date;
     private long timemillis;
